requirements.txt
click==8.0.4
dataclasses==0.8
Flask==2.0.3
Flask-MySQLdb==1.0.1
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mysqlclient==2.1.1
python-dotenv==0.20.0
typing_extensions==4.1.1
Werkzeug==2.0.3
zipp==3.6.0
